requirements.txt
python-telegram-bot==21.7
SQLAlchemy==1.4.23
Pillow>=9.0.0
qrcode==6.1
opencv-python>=4.1.2
numpy>=1.18.5
pdf2image==1.16.3
reportlab==4.0.7
poppler-utils==0.1.0
python-dotenv==0.19.0
ultralytics>=8.0.0
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
pyyaml>=5.3.1
seaborn>=0.11.0
pandas>=1.1.4